# page/openApi/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/authors":{"post":{"tags":["authors"],"summary":"Create Author","description":"著者登録 API エンドポイント\n\nArgs:\n    author_create_param: 登録する著者の入力 DTO\n    session: FastAPI が管理する SQLAlchemy セッション\n    service: 実処理となる関数\n\nReturns:\n    登録された著者のレスポンス DTO","operationId":"create_author_authors_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorCreateParam"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books":{"get":{"tags":["books"],"summary":"Find Books","description":"全件の書籍検索 API エンドポイント\n\nArgs:\n    session: FastAPI が管理する SQLAlchemy セッション\n    service: 実処理となる関数\n\nReturns:\n    すべての書籍のレスポンス DTO","operationId":"find_books_books_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BookResponse"},"type":"array","title":"Response Find Books Books Get"}}}}}},"post":{"tags":["books"],"summary":"Create Book","description":"書籍登録 API エンドポイント\n\nArgs:\n    book_create_param: 登録する書籍の入力 DTO\n    session: FastAPI が管理する SQLAlchemy セッション\n    service: 実処理となる関数\n\nReturns:\n    登録された書籍のレスポンス DTO","operationId":"create_book_books_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookCreateParam"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookResponse"}}}},"404":{"description":"Author not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundAuthorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/books/{book_id}":{"get":{"tags":["books"],"summary":"Find Book","description":"idが一致する書籍検索 API エンドポイント\n\nArgs:\n    book_id: 取得対象とする書籍の一意な識別子\n    session: FastAPI が管理する SQLAlchemy セッション\n    service: 実処理となる関数\n\nReturns:\n    条件に一致する書籍のレスポンス DTO","operationId":"find_book_books__book_id__get","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","description":"取得対象とする書籍の一意な識別子","title":"Book Id"},"description":"取得対象とする書籍の一意な識別子"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookResponse"}}}},"404":{"description":"Book not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundBookResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["books"],"summary":"Delete Book","description":"書籍削除 API エンドポイント\n\nArgs:\n    book_id: 削除対象とする書籍の一意な識別子\n    session: FastAPI が管理する SQLAlchemy セッション\n    service: 実処理となる関数","operationId":"delete_book_books__book_id__delete","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","description":"削除対象とする書籍の一意な識別子","title":"Book Id"},"description":"削除対象とする書籍の一意な識別子"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Book not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundBookResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Author":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"著者の一意な識別子","examples":["xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx"]},"name":{"type":"string","title":"Name","description":"著者名","examples":["著者の氏名"]}},"type":"object","required":["id","name"],"title":"Author"},"AuthorCreateParam":{"properties":{"name":{"type":"string","maxLength":50,"title":"Name","description":"著者名。最大 50 文字。","examples":["著者の氏名"]}},"additionalProperties":false,"type":"object","required":["name"],"title":"AuthorCreateParam","description":"著者作成用の入力 DTO"},"AuthorResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"著者の一意な識別子","examples":["xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx"]},"name":{"type":"string","title":"Name","description":"著者名","examples":["著者の氏名"]}},"type":"object","required":["id","name"],"title":"AuthorResponse","description":"著者レスポンス DTO"},"BookCreateParam":{"properties":{"title":{"type":"string","maxLength":100,"title":"Title","description":"書籍名。最大 100 文字。","examples":["著書の名前"]},"author_id":{"type":"string","format":"uuid","title":"Author Id","description":"著者ID。既存の著者IDを指定する必要があります。","examples":["xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx"]}},"additionalProperties":false,"type":"object","required":["title","author_id"],"title":"BookCreateParam","description":"書籍作成用の入力 DTO"},"BookResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"著書の一意な識別子","examples":["xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx"]},"title":{"type":"string","title":"Title","description":"書籍のタイトル","examples":["著書の名前"]},"author":{"$ref":"#/components/schemas/Author","description":"書籍の著者情報"}},"type":"object","required":["id","title","author"],"title":"BookResponse","description":"書籍レスポンス DTO"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NotFoundAuthorResponse":{"properties":{"message":{"type":"string","title":"Message","description":"メッセージ","examples":["著者が見つかりませんでした [id: xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx]"]},"author_id":{"type":"string","format":"uuid","title":"Author Id","description":"著者の一意な識別子","examples":["xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx"]}},"type":"object","required":["message","author_id"],"title":"NotFoundAuthorResponse","description":"著者未検出時のエラーレスポンス DTO"},"NotFoundBookResponse":{"properties":{"message":{"type":"string","title":"Message","description":"メッセージ","examples":["書籍が見つかりませんでした [id: xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx]"]},"book_id":{"type":"string","format":"uuid","title":"Book Id","description":"書籍の一意な識別子","examples":["xxxxxxxx-xxxx-xxxx-Nxxx-xxxxxxxxxxxx"]}},"type":"object","required":["message","book_id"],"title":"NotFoundBookResponse","description":"書籍未検出時のエラーレスポンス DTO"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}